% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{abs_d_ppvnpv}
\alias{abs_d_ppvnpv}
\title{Calculate the absolute difference of positive and negative predictive value}
\usage{
abs_d_ppvnpv(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the absolute difference of positive predictive value (PPV) and
negative predictive value (NPV)
from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
abs_d_ppvnpv(10, 5, 20, 10)
}
