% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{abs_d_sesp}
\alias{abs_d_sesp}
\title{Calculate absolute difference of sensitivity and specificity
Calculate the absolute difference of sensitivity and specificity
from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.}
\usage{
abs_d_sesp(tp, fp, tn, fn)
}
\arguments{
\item{tp}{(numeric) number of true positives}

\item{fp}{(numeric) number of false positives}

\item{tn}{(numeric) number of true negatives}

\item{fn}{(numeric) number of false negatives}
}
\description{
Calculate absolute difference of sensitivity and specificity
Calculate the absolute difference of sensitivity and specificity
from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
}
\examples{
abs_d_sesp(10, 5, 20, 10)
}

