% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{assume_direction_pos_class}
\alias{assume_direction_pos_class}
\title{Make assumptions for direction, pos_class and neg_class if necessary:}
\usage{
assume_direction_pos_class(x, class, pos_class, neg_class, direction, na.rm)
}
\description{
Make assumptions for direction, pos_class and neg_class if necessary:
}

