% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc}
\alias{auc}
\title{Calculate AUC from a roc_cutpointr object}
\source{
Forked from the AUC package
}
\usage{
auc(roc_cutpointr)
}
\arguments{
\item{roc_cutpointr}{A roc_cutpointr object resulting from the roc() function.}
}
\description{
Calculate the area under the ROC curve using the trapezoidal rule.
}
