% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cutpoint}
\alias{cutpoint}
\alias{cutpoints}
\title{Extract the cutpoints from a ROC curve generated by cutpointr}
\usage{
cutpoint(x, ...)

cutpoints(x, ...)
}
\arguments{
\item{x}{A roc_cutpointr object.}

\item{...}{Further arguments.}
}
\description{
This is a utility function for extracting the cutpoints from a \code{roc_cutpointr}
object. Mainly useful in conjunction with the \code{plot_cutpointr} function if
cutpoints are to be plotted on the x-axis.
}
\examples{
oc <- cutpointr(suicide, dsi, suicide, gender)
plot_cutpointr(oc, cutpoint, accuracy)

}
