% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_metric.R
\name{cutpoint_knots}
\alias{cutpoint_knots}
\title{Calculate number of knots to use in spline smoothing}
\usage{
cutpoint_knots(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{(character) The name of the predictor variable}
}
\description{
This function implements a rule of thumb for picking a number of knots
when using smoothing splines for smoothing a function of metric values per
cutpoint value.
}
\details{
The number of knots is equal to ceiling(0.1 * log(n) / n_cut) * n_cut) + 1
where n is the number of observations and n_cut is the number of unique
cutpoints found in the data.
}
\examples{
cutpoint_knots(suicide, "dsi")
}
