% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpointr.R
\name{cutpointr_}
\alias{cutpointr_}
\title{The standard evaluation version of cutpointr}
\usage{
cutpointr_(data, x, class, subgroup = NULL, method = "maximize_metric",
  metric = "sum_sens_spec", pos_class = NULL, neg_class = NULL,
  direction = NULL, boot_runs = 0, use_midpoints = FALSE, na.rm = FALSE,
  allowParallel = FALSE, ...)
}
\arguments{
\item{data}{A data frame or tibble in which the columns that are given in x,
class and possibly subgroup can be found.}

\item{x}{The variable name with quotation marks to be used for
classification, e.g. predictions or test values.}

\item{class}{The variable name with quotation marks indicating class membership.}

\item{subgroup}{The variable name with quotation marks
of an additional covariate that identifies subgroups. Separate
optimal cutpoints will be determined by group. Numeric, character and factor are
allowed. Also expressions like z > 10 are possible.}

\item{method}{(character) A function for determining cutpoints. Can
be user supplied or use some of the built in methods. See details.}

\item{metric}{(character) The function for computing a metric when using
maximize_metric or minimize_metric as method and and for the
out-of-bag values during bootstrapping. A way of internally validating the performance.
User defined functions can be supplied, see details.}

\item{pos_class}{(optional) The value of class that indicates the positive class.}

\item{neg_class}{(optional) The value of class that indicates the negative class.}

\item{direction}{(character, optional) Use ">=" or "<=" to indicate whether x
is supposed to be larger or smaller for the positive class.}

\item{boot_runs}{(numerical) If positive, this number of bootstrap samples
will be used to assess the variability and the out-of-sample performance.}

\item{use_midpoints}{(logical) If TRUE (default FALSE) the returned optimal
cutpoint will be the mean of the optimal cutpoint and the next highest
observation (for direction = ">") or the next lowest observation
(for direction = "<").}

\item{na.rm}{(logical) Set to TRUE (default FALSE) to keep only complete
cases of x, class and subgroup (if specified). Missing values with
na.rm = FALSE will raise an error.}

\item{allowParallel}{(logical) If TRUE, the bootstrapping will be parallelized
using foreach. A local cluster, for example, should have been started manually
beforehand.}

\item{...}{Further optional arguments that will be passed to optcut_func.}
}
\description{
This function is equivalent to cutpointr but takes only quoted arguments
for x, class, subgroup, method and metric. This function is suitable for
programming. For details on cutpointr see help("cutpointr").
}
