% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_opt_cut}
\alias{extract_opt_cut}
\title{Insert midpoints between neighboring values after sorting, usually for cutpoints}
\usage{
extract_opt_cut(df)
}
\description{
Insert midpoints between neighboring values after sorting, usually for cutpoints
}

