% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{F1_score}
\alias{F1_score}
\title{Calculate the F1-score from a 2x2 confusion matrix}
\usage{
F1_score(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the F1-score from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
F1_score(10, 5, 20, 10)
}
