% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{false_omission_rate}
\alias{false_omission_rate}
\alias{false_discovery_rate}
\title{Calculate the false omission and false discovery rate}
\usage{
false_omission_rate(tp, fp, tn, fn, ...)

false_discovery_rate(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the false omission rate or false discovery rate
from true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
false_omission_rate = fn / (tn + fn) = 1 - npv
false_discovery_rate = fp / (tp + fp) = 1 - ppv
}
\examples{
false_omission_rate(10, 5, 20, 10)
false_omission_rate(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
