% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_cutoffs.R
\name{gather_cutoffs}
\alias{gather_cutoffs}
\title{Calculate a metric function over given possible cutpoints}
\usage{
gather_cutoffs(data, x, class, candidate_cuts = unique(data[, x]),
  metric_func = NULL, pos_class, neg_class, direction = ">", ...)
}
\arguments{
\item{data}{A data frame or tbl_df}

\item{x}{(character) The column name of the predictor variable}

\item{class}{(character) The column name of the class (outcome) variable}

\item{candidate_cuts}{(numeric vector) Optionally, the cutoffs to be supplied
to metric_func can be specified. By default all unique values in x will be used.}

\item{metric_func}{(function) A function that takes one or more of the following
arguments:
\itemize{
 \item preds
 \item obs
 \item pos_class
 \item ...
}
The function should return a single numeric value.}

\item{pos_class}{The label of the positive class. For example, if the class
column in data is numeric, this should be numeric, too.}

\item{neg_class}{(optional) The label of the negative class. If not given,
if will be assumed as the value in the class column that is not pos_class.
Specifying neg_class is faster for larger data.}

\item{direction}{(character) This should be ">" (default) or "<" to specify
whether x values larger or smaller than a cutoff indicate the positive class.}

\item{...}{Further optional arguments that will be passed to metric_func}
}
\value{
A data frame with columns "cutpoint" and "metric"
}
\description{
A metric function, e.g. a function to calculate accuracy, has to be supplied
that will be applied to classifications of the data at every possible cutpoint,
i.e. all values in candidate_cuts.
}

