% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{inf_to_candidate_cuts}
\alias{inf_to_candidate_cuts}
\title{Add Inf or -Inf to candidate_cuts depending on direction}
\usage{
inf_to_candidate_cuts(candidate_cuts, direction)
}
\description{
Add Inf or -Inf to candidate_cuts depending on direction
}

