% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_midpoints}
\alias{insert_midpoints}
\title{Insert midpoints between neighboring values after sorting, usually for cutpoints}
\usage{
insert_midpoints(x)
}
\description{
Insert midpoints between neighboring values after sorting, usually for cutpoints
}

