% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{kappa}
\alias{kappa}
\title{Calculate the Kappa metric
Calculate the Kappa metric from the elements of a 2x2 confusion matrix,
that is true positives, false positives, true negatives and false negatives.}
\usage{
kappa(tp, fp, tn, fn)
}
\arguments{
\item{tp}{(numeric) number of true positives}

\item{fp}{(numeric) number of false positives}

\item{tn}{(numeric) number of true negatives}

\item{fn}{(numeric) number of false negatives}
}
\description{
Calculate the Kappa metric
Calculate the Kappa metric from the elements of a 2x2 confusion matrix,
that is true positives, false positives, true negatives and false negatives.
}
\examples{
kappa(10, 5, 20, 10)
}

