% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{log_loss}
\alias{log_loss}
\title{Calculate the Log Loss}
\usage{
log_loss(tp, fp, tn, fn, eps = 1e-15, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{eps}{Log loss is undefined for p=0 or p=1,
so probabilities are clipped to max(eps, min(1 - eps, p)).}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the Log Loss from true positives, false positives,
true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
Log_Loss = -(y log(p) + (1 - y) log(1 - p)), but
in the cases of thresholded values, p is either 0 or 1.
}
\examples{
log_loss(10, 5, 20, 10)
log_loss(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
