% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_metric.R
\name{minimize_loess_metric}
\alias{minimize_loess_metric}
\title{Minimize a metric function in binary classification after LOESS smoothing}
\source{
Xiao-Feng Wang (2010). fANCOVA: Nonparametric Analysis of Covariance.
 https://CRAN.R-project.org/package=fANCOVA
}
\usage{
minimize_loess_metric(data, x, class, metric_func = youden,
  pos_class = NULL, neg_class = NULL, direction, ...)
}
\arguments{
\item{data}{A data frame or tibble in which the columns that are given in x
and class can be found.}

\item{x}{(character) The variable name to be used for classification,
e.g. predictions or test values.}

\item{class}{(character) The variable name indicating class membership.}

\item{metric_func}{(function) A function that computes a single number
metric to be maximized. See description.}

\item{pos_class}{The value of class that indicates the positive class.}

\item{neg_class}{The value of class that indicates the negative class.}

\item{direction}{(character) Use ">=" or "<=" to select whether an x value
>= or <= the cutoff predicts the positive class.}

\item{...}{Further arguments that will be passed to metric_func and
additional arguments for the LOESS smoother:
\itemize{
 \item criterion: the criterion for automatic smoothing parameter selection:
 "aicc" denotes bias-corrected AIC criterion, "gcv" denotes generalized
 cross-validation.
 \item degree: the degree of the local polynomials to be used. It can be
 0, 1 or 2.
 \item family: if "gaussian" fitting is by least-squares, and if "symmetric"
 a re-descending M estimator is used with Tukey's biweight function.
 \item user.span: the user-defined parameter which controls the degree of
 smoothing.
}}
}
\value{
A tibble with the columns optimal_cutpoint, the corresponding metric
value and roccurve, a nested tibble that includes all possible cutoffs
and the corresponding numbers of true and false positives / negatives and
all corresponding metric values.
}
\description{
Given a function for computing a metric in metric_func, this function
smoothes the function of metric value against cutpoint using LOESS. Then, it
minimizes the metric by selecting an optimal cutpoint. For further details
on the LOESS smoothing see ?fANCOVA::loess.as.
The metric function should accept the following inputs:
\itemize{
 \item tp: vector of number of true positives
 \item fp: vector of number of false positives
 \item tn: vector of number of true negatives
 \item fn: vector of number of false negatives
}
}
\details{
The above inputs are arrived at by using all unique values in x, Inf, and
-Inf as possible cutpoints for classifying the variable in class.
}
