% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{misclassification_cost}
\alias{misclassification_cost}
\title{Calculate the misclassification cost from a 2x2 confusion matrix}
\usage{
misclassification_cost(tp, fp, tn, fn, cost_fp = 1, cost_fn = 1, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{cost_fp}{(numeric) the cost of a false positive}

\item{cost_fn}{(numeric) the cost of a false negative}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the misclassification cost from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The misclassification cost is equal to cost_fp * fp + cost_fn * fn.
The inputs must be vectors of equal length.
}
\examples{
misclassification_cost(10, 5, 20, 10, cost_fp = 1, cost_fn = 5)
}
