% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_equalsesp.R
\name{oc_equalsesp.default}
\alias{oc_equalsesp.default}
\title{Find optimal cutpoint by minimizing the difference between sensitivity and specificity}
\usage{
\method{oc_equalsesp}{default}(data, x, class, candidate_cuts = NULL,
  pos_class = NULL, neg_class = NULL, direction = NULL)
}
\description{
Find optimal cutpoint by minimizing the difference between sensitivity and specificity
}

