% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_kappa.R
\name{oc_kappa_simple}
\alias{oc_kappa_simple}
\title{Find optimal cutpoint to maximize Kappa}
\usage{
oc_kappa_simple(x, class, candidate_cuts = unique(x), pos_class = NULL,
  neg_class = NULL, direction = ">")
}
\arguments{
\item{x}{The variable name (with or without quotation marks) to be used for classification, e.g. predictions or test values.}

\item{class}{The variable name (with or without quotation marks) indicating class membership.}

\item{pos_class}{(optional) The value of class that indicates the positive class}

\item{neg_class}{(optional) The value of class that indicates the negative class}

\item{direction}{(character, optional) Use ">" or "<" to indicate whether x
is supposed to be larger or smaller for the positive class.}
}
\description{
Find optimal cutpoint to maximize Kappa
}

