% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_manual.R
\name{oc_manual}
\alias{oc_manual}
\title{Set a manual cutpoint for use with cutpointr}
\usage{
oc_manual(cutpoint, ...)
}
\description{
Mainly useful if bootstrap estimates of the out-of-bag performance of a
given cutpoint are desired, e.g. taking a cutpoint value from the literature.
}

