% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_youden.R
\name{oc_youden}
\alias{oc_youden}
\title{Determine optimal cutpoint by maximizing the sum of sensitivity and specificity}
\usage{
oc_youden(data, x, class, pos_class = NULL, neg_class = NULL,
  direction = ">")
}
\arguments{
\item{data}{A data frame or tibble in which the columns that are given in x, class and possibly subgroup can be found}

\item{x}{The variable name (with or without quotation marks) to be used for classification, e.g. predictions or test values.}

\item{class}{The variable name (with or without quotation marks) indicating class membership.}

\item{pos_class}{(optional) The value of class that indicates the positive class}

\item{neg_class}{(optional) The value of class that indicates the negative class}

\item{direction}{(character, optional) Use ">" or "<" to indicate whether x
is supposed to be larger or smaller for the positive class.}
}
\value{
A data frame with one row, the column optimal_cutpoint and the column
youden
}
\description{
This function takes bivariate data of predictions or test values and class
labels and determines an optimal cutpoint to empirically maximize
sensitivity + specificity - 1 (the J-Index or Youden-Index).
}
\examples{
library(OptimalCutpoints)
data(elas)
oc_youden(elas, "elas", "status", pos_class = 1, neg_class = 0)
}

