% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_youden_normal.R
\name{oc_youden_normal}
\alias{oc_youden_normal}
\title{Determine an optimal cutpoint for the Youden-Index assuming normal distributions}
\usage{
oc_youden_normal(data, x, class, pos_class = NULL, neg_class = NULL,
  direction, ...)
}
\description{
Determine an optimal cutpoint for the Youden-Index assuming normal distributions
}

