% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Calculate the odds ratio}
\usage{
odds_ratio(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the (diagnostic) odds ratio from
true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
odds_ratio = (tp / fp) / (fn / tn) \cr
}
\examples{
odds_ratio(10, 5, 20, 10)
odds_ratio(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
