% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optcut_emp_eqsensspec.R
\name{optcut_emp_eqsensspec.default}
\alias{optcut_emp_eqsensspec.default}
\title{Find optimal cutpoint by minimizing the difference between sensitivity and specificity}
\usage{
\method{optcut_emp_eqsensspec}{default}(x, class, candidate_cuts = unique(x),
  pos_class = NULL, higher = NULL)
}
\description{
Find optimal cutpoint by minimizing the difference between sensitivity and specificity
}

