% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optcut_emp_eqsensspec.R
\name{optcut_emp_eqsensspec}
\alias{optcut_emp_eqsensspec}
\title{Find optimal cutpoint by minimizing the difference between sensitivity and specificity}
\usage{
optcut_emp_eqsensspec(...)
}
\description{
Find optimal cutpoint by minimizing the difference between sensitivity and specificity
}

