% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{p_chisquared}
\alias{p_chisquared}
\title{Calculate the p-value of a chi-squared test}
\usage{
p_chisquared(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the p-value of a chi-squared test from
true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
p_chisquared(10, 5, 20, 10)
p_chisquared(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
