% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multi_cutpointr.R
\name{plot.multi_cutpointr}
\alias{plot.multi_cutpointr}
\title{Plotting multi_cutpointr objects is currently not supported}
\usage{
\method{plot}{multi_cutpointr}(x, ...)
}
\arguments{
\item{x}{A multi_cutpointr object.}

\item{...}{Further arguments.}
}
\description{
You can try plotting the data manually instead.
}
