% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precision_recall.R
\name{plot_precision_recall}
\alias{plot_precision_recall}
\title{Precision recall plot from a cutpointr object}
\usage{
plot_precision_recall(x, display_cutpoint = TRUE, ...)
}
\arguments{
\item{x}{A cutpointr object.}

\item{display_cutpoint}{(logical) Whether or not to display the optimal
cutpoint as a dot on the precision recall curve.}

\item{...}{Additional arguments (unused).}
}
\description{
Given a \code{cutpointr} object this function plots the precision recall curve(s)
per subgroup, if given.
}
\examples{
library(cutpointr)

## Optimal cutpoint for dsi
data(suicide)
opt_cut <- cutpointr(suicide, dsi, suicide)
plot_precision_recall(opt_cut)
}
