% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve from a cutpointr object}
\usage{
plot_roc(x, ...)
}
\arguments{
\item{x}{A cutpointr object}

\item{...}{Additional arguments (unused)}
}
\description{
Given a cutpointr object this function plots the ROC curve(s)
per subgroup, if given.
}

