% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x.R
\name{plot_x}
\alias{plot_x}
\title{Plot the distribution of the independent variable and the optimal cutpoints}
\usage{
plot_x(x, display_cutpoint = TRUE, ...)
}
\arguments{
\item{x}{A cutpointr object.}

\item{display_cutpoint}{(logical) Whether or not to display the optimal
cutpoint as a vertical line.}

\item{...}{Additional arguments (unused).}
}
\description{
Given a cutpointr object this function plots the distribution(s) of the
independent variable(s) and the respective cutpoints.
}
