% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plr}
\alias{plr}
\alias{nlr}
\title{Calculate the positive or negative likelihood ratio}
\usage{
plr(tp, fp, tn, fn, ...)

nlr(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the positive or negative likelihood ratio
from true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
plr = tpr / fpr \cr
nlr = fnr / tnr \cr
}
\examples{
plr(10, 5, 20, 10)
plr(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
