% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cutpointr.R
\name{print.cutpointr}
\alias{print.cutpointr}
\title{Print cutpointr objects}
\source{
Kirill Müller and Hadley Wickham (2017). tibble: Simple Data Frames.
 https://CRAN.R-project.org/package=tibble
}
\usage{
\method{print}{cutpointr}(x, n = Inf, ...)
}
\arguments{
\item{x}{a cutpointr object.}

\item{n}{number of rows to print.}

\item{...}{further arguements.}
}
\description{
Prints the cutpointr object like a tbl_df.
With versions of tibble later than 1.3.0 printing tibbles with width = Inf
does not wrap columns nicely so the default has been changed to width = NULL.
}
