% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multi_cutpointr.R
\name{print.multi_cutpointr}
\alias{print.multi_cutpointr}
\title{Print multi_cutpointr objects}
\source{
Kirill Müller and Hadley Wickham (2017). tibble: Simple Data Frames.
 https://CRAN.R-project.org/package=tibble
}
\usage{
\method{print}{multi_cutpointr}(x, n = Inf, ...)
}
\arguments{
\item{x}{a multi_cutpointr object.}

\item{n}{number of rows to print.}

\item{...}{further arguements.}
}
\description{
Prints the multi_cutpointr object with infinite width like a tbl_df.
}
