% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate_nodal}
\alias{prostate_nodal}
\title{Nodal involvement and acid phosphatase levels in 53 prostate cancer patients}
\format{A data frame with 53 rows and 2 variables:
\describe{
  \item{acid_phosphatase}{(numeric) Blood serum level of acid phosphatase}
  \item{nodal_inolvement}{(logical) Whether neighboring lymph nodes were involved}
}}
\source{
Le CT (2006). A solution for the most basic optimization problem associated with an ROC curve. Statistical methods in medical research 15: 571–584
}
\usage{
prostate_nodal
}
\description{
Prostatic acid phosphatase (PAP) emerged as the first clinically
useful tumor marker in the 1940s and 1950s. This data set contains the
serum levels of acid phosphatase of 53 patients that were confirmed to
have prostate cancer and whether the neighboring lymph nodes were involved.
}
\keyword{datasets}
