% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{risk_ratio}
\alias{risk_ratio}
\title{Calculate the risk ratio (relative risk)}
\usage{
risk_ratio(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the risk ratio (or relative risk) from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The risk ratio is defined as: (TP / FP) / (TN / FN)
The inputs must be vectors of equal length.
}
\examples{
risk_ratio(10, 5, 20, 10)
}
