% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Calculate a ROC curve}
\usage{
roc(data, x, class, pos_class, neg_class, direction = ">=")
}
\arguments{
\item{data}{A data frame or matrix. Will be converted to a data frame.}

\item{x}{(charaacter) The numeric independent (predictor) variable.}

\item{class}{(character) A binary vector of outcome values.}

\item{pos_class}{The value of 'class' that represents the positive cases.}

\item{neg_class}{The value of 'class' that represents the negative cases.}

\item{direction}{(character) One of ">=" or "<=". Specifies if the positive
class is associated with higher values of x (default).}
}
\value{
A data frame with the columns x.sorted, tp, fp, tn, fn, tpr, tnr, fpr,
and fnr.
}
\description{
Given a data frame with a numeric predictor variable and a binary outcome
variable this function returns a data frame that includes all elements of
a confusion matrix for every unique value of the predictor variable.
Additionally, the TPR, FPR, TNR and FNR are returned.
}
\examples{
## First two classes of the iris data
dat <- iris[1:100, ]
roc(data = dat, x = "Petal.Width", class = "Species",
pos_class = "versicolor", neg_class = "setosa", direction = ">=")
}

