% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate sensitivity}
\usage{
sensitivity(tp, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate sensitivity from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
sensitivity(10, 5, 20, 10)
}
