% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{specificity}
\alias{specificity}
\title{Calculate specificity}
\usage{
specificity(fp, tn, ...)
}
\arguments{
\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate specificity from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
specificity(10, 5, 20, 10)
}
