% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sum_ppv_npv}
\alias{sum_ppv_npv}
\title{Calculate the sum of positive and negative predictive value}
\usage{
sum_ppv_npv(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the sum of positive predictive value (PPV) and
negative predictive value (NPV)
from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
sum_ppv_npv(10, 5, 20, 10)
}
