% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.cutpointr}
\alias{tidy.cutpointr}
\title{Extract the optimal cutpoint from a cutpointr object or a cp_result object or a data.frame}
\usage{
tidy.cutpointr(data, maximize = T)
}
\arguments{
\item{maximize}{(logical) Whether to maximize or minimize the metric.
Default TRUE.}

\item{object}{(cutpointr) A data frame with the column "cutpoint" as the first
column and a second column of metric values.}

\item{break_ties}{(logical) If TRUE (default) the function will
return the median cutoff of all cutoff values that maximize the metric.
Otherwise the first (lowest) cutoff that maximizes the metric will be returned.}
}
\description{
In the spirit of broom::tidy. Works on any data.frame that has two columns.
The expected column name for the cutpoints is "cutpoint". The other one will
be treated as the metric to be maximized (or minimized). If no column named
"cutpoint" is found the first column will be treated as the column of cutpoints.
}

