% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{total_utility}
\alias{total_utility}
\title{Calculate the total utility from a 2x2 confusion matrix}
\usage{
total_utility(tp, fp, tn, fn, utility_tp = 1, utility_tn = 1, cost_fp = 1,
  cost_fn = 1, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{utility_tp}{(numeric) the utility of a true positive}

\item{utility_tn}{(numeric) the utility of a true negative}

\item{cost_fp}{(numeric) the cost of a false positive}

\item{cost_fn}{(numeric) the cost of a false negative}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the total benefit from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The total benefit is equal to
utility_tp * tp + utility_tn * tn - cost_fp * fp - cost_fn * fn.
The inputs must be vectors of equal length.
}
\examples{
total_utility(10, 5, 20, 10, utility_tp = 3, utility_tn = 3, cost_fp = 1, cost_fn = 5)
}
