% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{tp}
\alias{tp}
\alias{tn}
\alias{fp}
\alias{fn}
\title{Extract number true / false positives / negatives}
\usage{
tp(tp, ...)

tn(tn, ...)

fp(fp, ...)

fn(fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{...}{for capturing additional arguments passed by method.}

\item{tn}{(numeric) number of true negatives.}

\item{fp}{(numeric) number of false positives.}

\item{fn}{(numeric) number of false negatives.}
}
\description{
Extract the number of true positives (tp), false positives (fp),
true negatives (tn), or false negatives (fn).
The inputs must be vectors of equal length. Mainly useful for \code{plot_cutpointr}.
}
\examples{
tp(10, 5, 20, 10)
tp(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
fp(10, 5, 20, 10)
tn(10, 5, 20, 10)
fn(10, 5, 20, 10)
}
