% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{tpr}
\alias{tpr}
\alias{fpr}
\alias{tnr}
\alias{fnr}
\title{Calculate true / false positive / negative rate}
\usage{
tpr(tp, fn, ...)

fpr(fp, tn, ...)

tnr(fp, tn, ...)

fnr(tp, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}
}
\description{
Calculate the true positive rate (tpr, equal to sensitivity),
the false positive rate (fpr), the true negative rate (tnr, equal to sensitivity),
or the false negative rate (fnr)
from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
tpr(10, 5, 20, 10)
}
