% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{youden}
\alias{youden}
\title{Calculate the Youden-Index}
\usage{
youden(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the Youden-Index (J-Index) from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
youden(10, 5, 20, 10)
}
