% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{abs_d_sens_spec}
\alias{abs_d_sens_spec}
\title{Calculate the absolute difference of sensitivity and specificity}
\usage{
abs_d_sens_spec(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the absolute difference of sensitivity and specificity
from true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr
\cr
sensitivity = tp / (tp + fn) \cr
specificity = tn / (tn + fp) \cr
abs_d_sens_spec = |sensitivity - specificity| \cr
}
\examples{
abs_d_sens_spec(10, 5, 20, 10)
abs_d_sens_spec(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
