% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpointr.R
\name{cutpointr.default}
\alias{cutpointr.default}
\title{Determine and evaluate optimal cutpoints}
\usage{
\method{cutpointr}{default}(data, x, class, group, pos_class = NULL,
  higher = NULL, optcut_func = optcut_emp_youden, insert_midpoints = F,
  only_integer_cuts = F, boot_runs = 0)
}
\description{
Determine and evaluate optimal cutpoints
}
\examples{
library(OptimalCutpoints)
data(elas)
opt_cut <- cutpointr(elas, elas, status, gender, pos_class = 1, boot_runs = 500)
opt_cut
plot(opt_cut)
}

