% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpointr.R
\name{cutpointr}
\alias{cutpointr}
\title{Determine and evaluate optimal cutpoints}
\usage{
cutpointr(x, ...)
}
\arguments{
\item{x}{(numeric vector) The variable to be used for classification, e.g. test values.}

\item{class}{(vector) class is a binary vector of values indicating class membership.}

\item{group}{(vector) An additional covariate that identifies subgroups. Separate
optimal cutpoints will be determined by group.}
}
\description{
Determine and evaluate optimal cutpoints
}

