% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_metric.R
\name{minimize_metric}
\alias{minimize_metric}
\title{Minimize a metric function in binary classification}
\usage{
minimize_metric(data, x, class, metric_func = youden, pos_class = NULL,
  neg_class = NULL, direction)
}
\arguments{
\item{data}{A data frame or tibble in which the columns that are given in x
and class can be found.}

\item{x}{(character) The variable name to be used for classification,
e.g. predictions or test values.}

\item{class}{(character) The variable name indicating class membership.}

\item{metric_func}{(function) A function that computes a single number
metric to be minimized. See description.}

\item{pos_class}{The value of class that indicates the positive class.}

\item{neg_class}{The value of class that indicates the negative class.}

\item{direction}{(character) Use ">=" or "<=" to select whether an x value
>= or <= the cutoff predicts the positive class.}
}
\value{
A tibble with the columns optimal_cutpoint, the corresponding metric
value and roccurve, a nested tibble that includes all possible cutoffs
and the corresponding numbers of true and false positives / negatives and
all corresponding metric values.
}
\description{
Given a function for computing a metric in metric_func, this function
minimizes that metric by selecting an optimal cutpoint.
The metric function should accept the following inputs:
\itemize{
 \item tp: vector of number of true positives
 \item fp: vector of number of false positives
 \item tn: vector of number of true negatives
 \item fn: vector of number of false negatives
}
}
\details{
The above inputs are arrived at by using all unique values in x, Inf, and
-Inf as possible cutoffs for classifying the variable in class.
}
