% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpointr.R
\name{multi_cutpointr}
\alias{multi_cutpointr}
\title{Calculate optimal cutpoints and further statistics for multiple predictors}
\usage{
multi_cutpointr(data, x = colnames(data)[colnames(data) != class], class,
  silent = FALSE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Character vector of predictor variables.}

\item{class}{The name of the outcome / independent variable.}

\item{silent}{Whether to display messages.}

\item{...}{Further arguments to be passed to cutpointr.}
}
\value{
A data frame.
}
\description{
Runs cutpointr_ over multiple predictor variables. By default, cutpointr
will be run using all columns in the data set as predictors except for the
variable in class.
}
\details{
The automatic determination of positive / negative classes and direction
will be carried out separately for every predictor variable. That way, if
direction and the classes are not specified, the reported AUC for every
variable will be >= 0.5. AUC may be < 0.5 in the case of subgroups as
direction is equal in every subgroup.
}
\examples{
library(cutpointr)

multi_cutpointr(suicide, x = c("age", "dsi"), class = "suicide",
                pos_class = "yes")

multi_cutpointr(suicide, x = c("age", "dsi"), class = "suicide",
                subgroup = "gender", pos_class = "yes")

}
