% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_OptimalCutpoints.R
\name{oc_OptimalCutpoints}
\alias{oc_OptimalCutpoints}
\title{A tidy wrapper for optimal.cutpoints}
\usage{
oc_OptimalCutpoints(data, x, class, oc_metric, neg_class, direction,
  break_ties = mean, ...)
}
\arguments{
\item{data}{A data frame or tibble in which the columns that are given in x
and class can be found.}

\item{x}{(character) The variable name to be used for classification,
e.g. predictions or test values.}

\item{class}{(character) The variable name indicating class membership.}

\item{oc_metric}{(character) See ?optimal.cutpoints for available methods.}

\item{neg_class}{The value of class that indicates the negative class.}

\item{direction}{(character) Use ">=" or "<=" to select whether an x value
>= or <= the cutoff predicts the positive class.}

\item{break_ties}{(function) A function to handle multiple optimal cutpoints,
if multiple optimal cutpoints are found.}

\item{...}{To capture further arguments that are always passed to the method
function by cutpointr. The cutpointr function passes data, x, class,
metric_func, direction, pos_class and neg_class to the method function.}
}
\value{
A data frame with one row, the column optimal_cutpoint and a second
column named after method that gives the obtained metric value.
}
\description{
Determine "optimal" cutpoints using optimal.cutpoints and its included methods
from the OptimalCutpoints package. The output will be a tidy data frame.of
the optimal cutpoint and the value of the metric, depending on methods, that
was obtained in-sample.
}
\examples{
data(suicide)
if (require(OptimalCutpoints)) {
  oc_OptimalCutpoints(suicide, "dsi", "suicide", oc_metric = "Youden",
  neg_class = "no", direction = ">=")
}
}

