% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_youden.R
\name{oc_youden}
\alias{oc_youden}
\title{Determine optimal cutpoint by maximizing the sum of sensitivity and specificity}
\usage{
oc_youden(data, x, class, candidate_cuts = unique(unlist(data[, x])),
  pos_class = NULL, neg_class = NULL, direction = NULL)
}
\arguments{
\item{data}{(data.frame) The dependent and independent variables}

\item{x}{(numeric) The values that will be used for splitting}

\item{class}{(character / factor / numeric) The dependent binary class variable}
}
\description{
Cutpoints are taken from the sample values (with the optional addition of midpoints
between these values) and the classification quality of these cupoints is
evaluated based on the chosen method.
}
\examples{
library(OptimalCutpoints)
data(elas)
oc_youden(elas$elas, elas$status, pos_class = 1)
}

