% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Calculate the odds ratio
Calculate the (diagnostic) odds ratio from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The odds ratio is defined as: (TP / FP) / (TN / FN)
The inputs must be vectors of equal length.}
\usage{
odds_ratio(tp, fp, tn, fn)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}
}
\description{
Calculate the odds ratio
Calculate the (diagnostic) odds ratio from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The odds ratio is defined as: (TP / FP) / (TN / FN)
The inputs must be vectors of equal length.
}
\examples{
odds_ratio(10, 5, 20, 10)
}
