% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optcut_emp_youden.R
\name{optcut_emp_youden}
\alias{optcut_emp_youden}
\alias{optcut_emp_youden.default}
\alias{optcut_emp_youden.formula}
\title{Determine optimal cutpoint by maximizing the sum of sensitivity and specificity}
\usage{
optcut_emp_youden(...)

\method{optcut_emp_youden}{default}(x, class, candidate_cuts = unique(x),
  pos_class = NULL, higher = NULL)

\method{optcut_emp_youden}{formula}(formula, data)
}
\arguments{
\item{x}{(numeric) The values that will be used for splitting}

\item{class}{(character / factor / numeric) The dependent binary class variable}

\item{formula}{(formula) Formula for splitting classes by some numeric vector, e.g. class ~ x.}

\item{data}{(data.frame) The dependent and independent variables}
}
\description{
Cutpoints are taken from the sample values (with the optional addition of midpoints
between these values) and the classification quality of these cupoints is
evaluated based on the chosen method.
}
\examples{
dat <- iris[1:100, ]
optcut_emp_youden(dat$Sepal.Width, class = dat$Species)
library(OptimalCutpoints)
data(elas)
optcut_emp_youden(elas$elas, elas$status, pos_class = 1)
}

