% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_boot.R
\name{plot_metric_boot}
\alias{plot_metric_boot}
\title{Plot the bootstrapped metric distribution}
\usage{
plot_metric_boot(x, ...)
}
\arguments{
\item{x}{A cutpointr object.}

\item{...}{Additional arguments (unused)}
}
\description{
Given a cutpointr object this function plots the bootstrapped metric distribution.
The metric depends on the function that was supplied to `metric` in the
call to cutpointr.
The cutpointr function has to be run with boot_runs` > 0 to enable bootstrapping.
}

