% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cutpointr.R
\name{print.cutpointr}
\alias{print.cutpointr}
\title{Print cutpointr objects}
\source{
Kirill Müller and Hadley Wickham (2017). tibble: Simple Data Frames.
 https://CRAN.R-project.org/package=tibble
}
\usage{
\method{print}{cutpointr}(x, width = 1000, n = 50, sigfig = 6, ...)
}
\arguments{
\item{x}{a cutpointr object.}

\item{width}{width of output.}

\item{n}{number of rows to print.}

\item{sigfig}{Number of significant digits to print. Temporarily
overrides options("pillar.sigfig").}

\item{...}{further arguments.}
}
\description{
Prints the \code{cutpointr} object with full width like a \code{tbl_df}.
}
