% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{prod_sens_spec}
\alias{prod_sens_spec}
\title{Calculate the product of sensitivity and specificity}
\usage{
prod_sens_spec(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the product of sensitivity and specificity from
true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
sensitivity = tp / (tp + fn) \cr
specificity = tn / (tn + fp) \cr
prod_sens_spec = sensitivity * specificity \cr
}
\examples{
prod_sens_spec(10, 5, 20, 10)
prod_sens_spec(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
