% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate sensitivity}
\usage{
sensitivity(tp, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate sensitivity from
true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
sensitivity = tp / (tp + fn) \cr
}
\examples{
sensitivity(10, 5, 20, 10)
sensitivity(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
