% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{specificity}
\alias{specificity}
\title{Calculate specificity}
\usage{
specificity(fp, tn, ...)
}
\arguments{
\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate specificity from true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
specificity = tn / (tn + fp) \cr
}
\examples{
specificity(10, 5, 20, 10)
specificity(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
