% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sum_ppv_npv}
\alias{sum_ppv_npv}
\title{Calculate the sum of positive and negative predictive value}
\usage{
sum_ppv_npv(tp, fp, tn, fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}

\item{...}{for capturing additional arguments passed by method.}
}
\description{
Calculate the sum of positive predictive value (PPV) and
negative predictive value (NPV) from
true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length. \cr \cr
ppv = tp / (tp + fp) \cr
npv = tn / (tn + fn) \cr
sum_ppv_npv = ppv + npv \cr
}
\examples{
sum_ppv_npv(10, 5, 20, 10)
sum_ppv_npv(c(10, 8), c(5, 7), c(20, 12), c(10, 18))
}
\seealso{
Other metric functions: \code{\link{F1_score}},
  \code{\link{abs_d_ppv_npv}},
  \code{\link{abs_d_sens_spec}}, \code{\link{accuracy}},
  \code{\link{cohens_kappa}}, \code{\link{cutpoint}},
  \code{\link{false_omission_rate}},
  \code{\link{metric_constrain}},
  \code{\link{misclassification_cost}}, \code{\link{npv}},
  \code{\link{odds_ratio}}, \code{\link{p_chisquared}},
  \code{\link{plr}}, \code{\link{ppv}},
  \code{\link{precision}}, \code{\link{prod_ppv_npv}},
  \code{\link{prod_sens_spec}}, \code{\link{recall}},
  \code{\link{risk_ratio}}, \code{\link{roc01}},
  \code{\link{sensitivity}}, \code{\link{specificity}},
  \code{\link{sum_sens_spec}}, \code{\link{total_utility}},
  \code{\link{tpr}}, \code{\link{tp}}, \code{\link{youden}}
}
\concept{metric functions}
