% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sum_sens_spec}
\alias{sum_sens_spec}
\title{Calculate the sum of sensitivity and specificity
Calculate the sum of sensitivity and specificity from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.}
\usage{
sum_sens_spec(tp, fp, tn, fn)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{fp}{(numeric) number of false positives.}

\item{tn}{(numeric) number of true negatives.}

\item{fn}{(numeric) number of false negatives.}
}
\description{
Calculate the sum of sensitivity and specificity
Calculate the sum of sensitivity and specificity from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
sum_sens_spec(10, 5, 20, 10)
}
