% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{tp}
\alias{tp}
\alias{tn}
\alias{fp}
\alias{fn}
\title{Extract number true / false positives / negatives}
\usage{
tp(tp, ...)

tn(tn, ...)

fp(fp, ...)

fn(fn, ...)
}
\arguments{
\item{tp}{(numeric) number of true positives.}

\item{...}{for capturing additional arguments passed by method.}

\item{tn}{(numeric) number of true negatives.}

\item{fp}{(numeric) number of false positives.}

\item{fn}{(numeric) number of false negatives.}
}
\description{
Extract the number of true positives (tp), false positives (fp),
true negatives (tn), or false negatives (fn)
from the elements of a confusion matrix,
that is true positives, false positives, true negatives and false negatives.
The inputs must be vectors of equal length.
}
\examples{
tp(10, 5, 20, 10)
fp(10, 5, 20, 10)
tn(10, 5, 20, 10)
fn(10, 5, 20, 10)
}
